[HWMETRICS-52] separate interface and impl classes
