+++ task-queue/src/main/java/org/hawkular/metrics/tasks/api/TaskType.java
@@ -14,13 +14,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.metrics.tasks;
+package org.hawkular.metrics.tasks.api;
 
 import java.util.Objects;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 
+import org.hawkular.metrics.tasks.impl.TaskImpl;
+
 /**
+ * Tasks are grouped by type, and execution order is also determined by type. If there are types A, B, C, then for
+ * a given time slice all tasks of type A are executed followed by tasks of type B and finally tasks of type C are
+ * executed. The task type specifies the granularity of partitioning of tasks in the database. It also provides a
+ * factory function that produces a function to execute tasks.
+ *
  * @author jsanda
  */
 public class TaskType {
@@ -33,6 +40,9 @@ public class TaskType {
 
     private int segmentOffsets;
 
+    /**
+     * The task type name which must be unique among task types.
+     */
     public String getName() {
         return name;
     }
@@ -42,6 +52,9 @@ public class TaskType {
         return this;
     }
 
+    /**
+     * A function that produces functions that carry out task execution.
+     */
     public Supplier<Consumer<Task>> getFactory() {
         return factory;
     }
@@ -51,6 +64,10 @@ public class TaskType {
         return this;
     }
 
+    /**
+     * The number of partitions in the database to use for storing tasks. The segment is determined by,
+     * <code>task.target.hashCode() % segments</code>
+     */
     public int getSegments() {
         return segments;
     }
@@ -60,6 +77,14 @@ public class TaskType {
         return this;
     }
 
+    /**
+     * Tasks are associated with leases. A client needs to acquire a lease before it can execute the tasks associated
+     * with the lease. This property specifies how many tasks in terms of segments are associated with a lease. Let's
+     * say that we have 100 segments and 10 segment offsets. This means that there are 10 task segments per lease.
+     * When a client acquires a lease, the client will then execute the tasks in each of those segments.
+     * <br/><br/>
+     * TODO Come up with a more descriptive name
+     */
     public int getSegmentOffsets() {
         return segmentOffsets;
     }
@@ -69,6 +94,15 @@ public class TaskType {
         return this;
     }
 
+    /**
+     * A factory method for creating tasks of this type.
+     *
+     * @param target Identifier or key of the entity or time series associated with the task
+     * @param source Identifiers or keys of the time series that provide input data for the task
+     * @param interval Specifies the frequency of execution in minutes
+     * @param window Specifies the amount of data to include in minutes
+     * @return A {@link Task}
+     */
     public Task createTask(String target, String source, int interval, int window) {
         return new TaskImpl(this, null, target, source, interval, window);
     }
