+++ task-queue/src/main/java/org/hawkular/metrics/tasks/impl/LeaseManager.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.metrics.tasks;
+package org.hawkular.metrics.tasks.impl;
 
 import static java.util.stream.Collectors.toList;
