+++ task-queue/src/main/java/org/hawkular/metrics/tasks/impl/Queries.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.metrics.tasks;
+package org.hawkular.metrics.tasks.impl;
 
 import com.datastax.driver.core.PreparedStatement;
 import com.datastax.driver.core.Session;
