+++ task-queue/src/main/java/org/hawkular/metrics/tasks/impl/TaskImpl.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.metrics.tasks;
+package org.hawkular.metrics.tasks.impl;
 
 import static org.joda.time.Duration.standardMinutes;
 
@@ -22,6 +22,8 @@ import java.util.Objects;
 import java.util.Set;
 
 import com.google.common.collect.ImmutableSet;
+import org.hawkular.metrics.tasks.api.TaskType;
+import org.hawkular.metrics.tasks.api.Task;
 import org.joda.time.DateTime;
 import org.joda.time.Duration;
